Removed "fake date"
A better way to handle plotting the daily records with the observed values was to just use a day of month column. This removes the ability to display the year of the record and the month date, year of the observed however. Need to figure out how to plot against one xtick, but display a different value on mouseover.